+++ cpp/src/barretenberg/common/msgpack_bin.cpp
@@ -4,17 +4,21 @@
 #include <barretenberg/common/net.hpp>
 #include <barretenberg/common/throw_or_abort.hpp>
 
-void msgpack_read_bin64(const msgpack::object& obj, unsigned long long* data, size_t size)
+namespace msgpack {
+
+void read_bin64(const msgpack::object& obj, uint64_t* data, size_t size)
 {
     if (obj.type != msgpack::type::BIN) {
         throw_or_abort("Wrong data type when unpacking bin64");
     }
-    if (obj.via.bin.size != size * sizeof(unsigned long long)) {
+    if (obj.via.bin.size != size * sizeof(uint64_t) {
         throw_or_abort("Wrong size data while unpacking bin64");
     }
 
-    unsigned long long* bin_data = (uint64_t*)obj.via.bin.ptr;
+    uint64_t* bin_data = (uint64_t*)obj.via.bin.ptr;
     for (size_t i = 0; i < size; i++) {
         data[i] = ntohll(bin_data[i]);
     }
 }
+
+} // namespace msgpack
